+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.0.0] 2021-03-10
+
+- New command **hardis:data:tree:export**
+- **scratch:create**: Import init data using .sfdx-hardis.yml `initDataRequests` property
+
 ## [2.0.0] 2021-03-09
 
 - New command **hardis:package:create** to create Managed and Unlocked packages